Fix precommit hook
In editorconfig `max_line_length = null` is invalid and results in an
error. Removing it corrects that. Additionally, the `lint-staged`
command was running `yarn test` on the staged js files individually,
resulting in it failing with a no test found message. Instead the tests
are run before the lint-staged command.