Rework aliases and platform for custom platforms
In aliases we put all the possible values into a map, respecting the
initial design of having the top level platform name checked first
then checking the aliases. More importantly this adds the ability to
specify custom platforms, more on that below.

In platforms we move the logic for each extension into a handler
function that takes the filename as an argument and either returns
the platform or false. More importantly, as above, we add the ability
to handle custom platforms. Custom platforms will be passed from the
config (not here, coming in next commit) and will be an object with
the platform name as the key and value a handler that takes the file
name and returns true if it's a match or false. If true is returned
then we return the custom platform name.

This commit also allows `mac`, `osx`, or `darwin` to be mapped to
`darwin`.