Centralize logic for getting latest for platform
This moves the logic that was previously duplicated in `download`
and `downloadPlatform` into `cache.getLatestForPlatform`.

Additionally, it adds the ability to set custom platforms. To set a
specify custom platforms you pass an object into the config as
`customPlatforms` where the key is the platform name and the value
is a function that takes the file name and returns true if that asset
should be used for the custom platform.

This also adds prioritization of .dmg files in the download handlers
for `darwin`. It does so by specifying `zip` as the filetype for
`darwin` update URLs for private repos and for public repos simply
getting the zip asset in the update handler.